Layering: Check for the correct owner in DetachArrayBuffer (#1112)
* Layering: Check for the correct owner in DetachArrayBuffer

This patch gives ArrayBuffers an "owner" internal slot. This slot is
used to throw a TypeError on inappropriate DetachArrayBuffer calls.

WebAssembly needs it to be an error to detach an ArrayBuffer (e.g.,
through postMessage) for ArrayBuffers which come from
WebAssembly.Memory objects. It should only be the Memory grow
operation which ever detaches this ArrayBuffer.

* Editorial: Clarifications based on code review