Editorial: Remove "Outside" wording for eval errors (#1245)
 - Be explicit about interleaving errors

It took me a few times over the specification to understand that the
headings and introductory text with descriptions like
"when a direct eval call occurs outside of any function"
were non-normative. Here's a confusing case:

new class { constructor() { function f() { eval("super()") } f(); } }

This example will throw a SyntaxError because the enclosing
ThisEnvironment when eval runs is not a derived constructor. However,
in the ordinary English sense of "outside", I'd say that the eval
call is not outside of a constructor.

This patch removes the use of the term "outside" when describing the
distribution of this error case. It puts the error text inline in
PerformEval, where it's easier to see the relationship with where
the error condition comes from.