Editorial: Eliminate the phrase "is called on" (#1266)
For each of Function.prototype.{apply,call,toString}, the preamble begins:
    "When the `foo` method is called on an object _func_"
but it's unclear what "X is called on Y" means.

(The phrase isn't used anywhere else in the spec.
The active version, "calls X on Y" is used a few times,
but usually X is an abstract op and Y is an argument passed to it,
which doesn't apply here.)

Rather than defining the phrase, remove it,
and add a step to the algorithm to define _func_
in conventional terms.