Normative: IsRegExp: make `Symbol.match` not be special
Either `IsRegExp` should check all 4 (soon 5) of the regex symbols, or
it should check none.

Currently this check is only used in 4 places: 3 String methods, to
throw a TypeError to indicate that they only take a string and not a
regex; and in the `RegExp` constructor to decide whether to extract
`.constructor`/`.source`/`.flags`, or to naively stringify.

This will only impact objects that do not use `extends RegExp`/`super`
to create a "RegExp-like" object.