Normative: Make non-writable prototype properties not prevent assigning to instance
TC39 is discussing making inherited non-writable properties overridable with Set,
in #1307. We have not yet assessed the web compatibility of the change or written
tests, so this PR is not yet ready to merge, but this PR gives some concreteness
to how the specification may be done, to help move the discussion forward.

This patch implements the approach described in https://github.com/tc39/ecma262/pull/1307#issuecomment-422808140