Bug: Add missing has operation to property operations
This adds direct language about "does not have an own property with key".
This was brought up due to legacy WebIDL usage of this language from
host behavior when implementing their own exotic objects. This has been
in use for years and cannot be removed and should not be changed, this
is just explicitly adding it to the list of operations.

The spec already requires this operation be done by implementations due
to language like "... does not have an own property with key ..."