Layering: Rename Module.Instantiate to Module.Link (#1312)
The terminology between WebAssembly and ES modules differs when it comes to
instantiation, which causes some confusion for the integration of the two
systems. Ideally, we wouldn't use the same term to mean two different things
between the specs.

After talking about it with multiple people, it seems like calling it "the
linking phase" might be a better name for this phase anyway. Developers
are often confused about what actually happens during instantiation and
why it's separated from evaluation. I believe "linking" more clearly
communicates the justification for this being its own phase.