Editorial: centralize strictness for function defs (#1563)
Formerly, for every different form of function definition,
we would ask whether:
    the function code for the function definition is strict mode code
and then pass the result to *FunctionCreate,
and from there to FunctionAllocate.

Instead, we can equivalently ask whether:
    the source text of the function definition's body is strict mode code.
The thing is, *this* question can be asked for all function definitions
in one place: FunctionInitialize.

So do that, and drop the _strict_ parameter from FunctionAllocate.