Editorial: Fix wording in Note for Arguments exotic objects
The prose for Arguments Exotic Object
says that it maps "array-indexed properties"
to formal parameter bindings.

And yet the Note refers to "integer-indexed data properties".

Now, "array-indexed" and "integer-indexed" aren't defined,
but "array index" and "integer index" are,
so this could be taken as a disagreement.
Change the Note to agree with the normative prose.

(This use of "integer-indexed" was introduced before "integer index"
had a specific definition, so I think it just didn't get cleaned up
when "integer index" was introduced.)