Editorial: Merge 'ExtendedAtom' and 'Atom'
Note that:

ExtendedAtom was only ever 'invoked' under [~U],
and when the merged production is invoked with [~U],
it exactly reproduces the RHSs of ExtendedAtom.

Atom was only ever invoked under [+U],
and when the merged production is invoked with [+U],
it reproduces the RHSs of former Atom
except for the placement of the PatternCharacter RHS,
but that's okay, because former Atom wasn't an order-disambiguated production.