Normative: `ValidateAndApplyPropertyDescriptor`: fix bug that bypasses "apply"
This bug was introduced in https://github.com/tc39/ecma262/pull/353/ -
it prevents the change from actually being applied.

I believe the confusion was trying to be consistent with the `return true`
above, but that one is in the "current is undefined" section, in which
the change is not applied anyways.