meta: make "new feature" issue template clearer (#1687)
Make this more clearer can help to prevent the 'close issues' from recurring.

Better is not assume that the error is always from the user who has not read or not interpret properly.

If the number of 'close issues' does not decrease, we will have data to say that the text is still bad and could improve even further.

To learn more: https://en.wikipedia.org/wiki/Copywriting