Editorial: fix bug/inconsistency by inserting "Element Type" (#1734)
When extracting an 'Element Size' from the TypedArray Constructors table,
we have two syntaxes:
- `the Element Size ... for _foo_`
- `the Element Size ... for Element Type _foo_`

With the first, _foo_ is the name of a constructor from column 1 of the table.

With the second, _foo_ is an element type from column 2 of the table.

However, there's one line that uses the first syntax,
but with the second semantics.
This commit changes it to use the second syntax.

(This bug/inconsistency goes back to ES6, and was brought to light by the changes from #1725)