Normative: allow non-BMP characters in capture group names (#1932)
This is a followup to #1869.
Per consensus, `/(?<𝒜>.)/` should be legal -
but the rules for parsing non-u patterns are such that the source text is parsed by treating each half of a surrogate pair as an individual code point,
so the rules for RegExpIdentifierStart and RegExpIdentifierPart need to be tweaked to allow surrogate pairs for that case.