Normative: Add machinery to track the incumbent settings object in HTML for Job callbacks (#2086)
In HTML, JS functions that are called asynchronously as part of jobs
(like Promise microtasks) are generally associated with a piece of state
called an incumbent settings object. The incumbent is stored at the time
when the callback is first passed to the API that is responsible for
scheduling the job (e.g. Promise.then), and restored when the callback
is called.

For symmetry, all WebIDL callbacks do this:
https://heycam.github.io/webidl/#idl-callback-interface

This PR adds the necessary host hook machinery to let HTML store and
restore the incumbent settings object. The default behavior is nop.