Normative: Re-resolve unresolvable references on the global in PutValue.
Currently, JSC, V8, SM, and XS all *do not throw* on the following code,
contra spec:

```javascript
"use strict";
undeclared = (this.undeclared = 42);
```

This has been a bug in implementations for a decade
(https://bugzilla.mozilla.org/show_bug.cgi?id=605515), and is arguably
implementation reality not only for the web but for most
implementations.

This changes the spec to match the implemented re-resolution behavior.

Also see
- https://github.com/tc39/ecma262/issues/467
- https://github.com/tc39/test262/issues/1964