Editorial: Drop [[ScriptOrModule]] slot for exotic built-ins (#2190)
A built-in function that is implemented as an exotic object
has no need for a [[ScriptOrModule]] internal slot.
It always has the value `*null*`,
and the only place it's used is in
[the [[Call]] and [[Construct]] methods for exotic built-ins](sec-built-in-function-objects-call-thisargument-argumentslist),
so we can just inline `*null*` there.

The remaining 3 appearances of [[ScriptOrModule]]
pertain only to ordinary functions.

(Setting [[ScriptOrModule]] for exotic built-ins
made more sense when the slot was introduced,
because of the way GetActiveScriptOrModule() was written,
but PR #556 changed it to basically the current situation.)