Editorial: Swap 2 alternatives of Term in Annex B (#2228)
In this annex, the order of a production's alternatives is significant:
"each alternative is considered only if previous production alternatives do not match"

So in the partial production
    Term :: Atom | Atom Quantifier
the `Atom Quantifier` alternative will be considered only if
the `Atom` alternative doesn't match.

But at any point where `Atom Quantifier` would match (e.g., in /x*/u),
`Atom` would necessarily also match, thus preventing `Atom Quantifier`
from ever being considered. (In such a case, the `Atom` match would
result in a syntax error.)

This is clearly not the intended behavior:
(a) There's no point having an alternative that can never be used.
(b) It contradicts the guarantee that none of the annex's extensions
    change the syntax of [+U] patterns.

To restore the main-body syntax of [+U] patterns, I believe it suffices
to swap the order of the two alternatives.