Editorial: Modernize overview paragraph comparing ES to other PLs
Closes #2266

References to other languages' specifications and papers are also
removed, as the editor group does not wish to keep an up-to-date
non-normative bibliography to other specifications (some of which are
withdrawn, like Scheme's). Additionally, web searching technology has
advanced sufficiently that finding information on the other PLs should
be easy.