Editorial: Drop "in original [key] insertion order" (#2330)
The form
    "For each [element] of [list], in original key insertion order"
(or "... original insertion order") is weird because it implies that
it's not enough to know the order of the elements in the List,
you also have to know the order that they were inserted into the List.
But in fact, that's not necessary: in every case where this form is used,
the insertion order is the *same* as the List order,
because insertions only happen via appending to the List.

Map.prototype.forEach:
  Iterates over _map_.[[MapData]],
  whose only inserts are appends, in Map.prototype.set.

Set.prototype.forEach:
  Iterates over _set_.[[SetData]],
  whose only inserts are appends, in Set.prototype.add.

TriggerPromiseReactions:
  Iterates over _reactions_, which is
  _promise_.[[PromiseFulfillReactions]] or
  _promise_.[[PromiseRejectReactions]],
  whose only inserts are appends, in PerformPromiseThen.