Editorial: Eliminate one use of SDO `CoveredFormalsList` (#2350)
Here in InstantiateArrowFunctionExpression,
I assert that we can replace the status quo's
    CoveredFormalsList of |ArrowParameters|
with just
    |ArrowParameters|
(and then eliminate the _parameters_ alias).

case analysis on the 2 forms of |ArrowParameters|:

case 1:
When |ArrowParameters| is an instance of
    ArrowParameters : BindingIdentifier
then CoveredFormalsList simply returns |ArrowParameters| itself,
so the above replacement is obviously correct.

case 2:
When |ArrowParameters| is an instance of
    ArrowParameters : CoverParenthesizedExpressionAndArrowParameterList
then CoveredFormalsList will return an instance of
    ArrowFormalParameters
(whose particulars are not important here),
which then gets passed to OrdinaryFunctionCreate's ParameterList.

But if you go to OrdinaryFunctionCreate and follow what happens
to ParameterList, you find that, ultimately, it's submitted to:
    - ExpectedArgumentCount
    - BoundNames
    - IsSimpleParameterList
    - ContainsExpression
    - IteratorBindingInitialization

And all of these have an explicit definition for:
    ArrowParameters : CoverParenthesizedExpressionAndArrowParameterList
of roughly the form:
    1. Let _formals_ be the CoveredFormalsList of |CoverParenthesizedExpressionAndArrowParameterList|.
    1. Return <SDO> of _formals_.

That is, with respect to the OrdinaryFunctionCreate's call-tree,
if you don't call CoveredFormalsList near the root,
it will be called near the leaves.

So in this case too, it's okay to perform the given replacement.