Editorial: Define `Evaluation` for BindingIdentifier (#2406)
PR #614 defined (in Annex B) an additional form for ForInOfStatement.
Performing ForInOfLoopEvaluation on it ends with a call to ForIn/OfBodyEvaluation,
with |BindingIdentifier| as first arg, passed to the _lhs_ param.
ForIn/OfBodyEvaluation (at step 6.g.i.1) evaluates this param,
but there's no definition of Evaluation for |BindingIdentifier|.
This PR adds the necessary definition.

Rather than add it to the annex that defines the additional form of ForInOfStatement,
this commit substitutes it for the (main body) definition of
`Evaluation` for `ForBinding : BindingIdentifier`,
allowing the latter to chain to the new definition,
since it has the same semantics.