Editorial: Remove <SP> and <NBSP> as introduced by <USP> (#2559)
U+0020 Space and U+00A0 No-Break Space belong to Space_Separator
category, so they are implicitly introduced into a right-hand side
of WhiteSpace by <USP> and need no repetition.

To avoid confusion of a reader, a reminder note is added.

---

Remove a note about Zs that duplicates the grammar

<USP> (all Space_Separator code points) is already listed as a
right-hand side of WhiteSpace, so there is no need to remind about it,
especially right before a declaration of WhiteSpace.