Editorial: Consistify prose for same-value properties (#2575)
Use consistent prose when the initial values of two intrinsic-properties are the same object.

In the sentences affected by this PR, refer to properties by just the final property name rather than by "full paths".

That is, change:
- `Number.parseFloat` to *"parseFloat"*
- `Number.parseInt` to *"parseInt"*
- %TypedArray%`.prototype.toString` to *"toString"*
- `Date.prototype.toGMTString` to *"toGMTString"*

This increases consistency among sentences of this particular form.

(But note that they're a small fraction of a large class of sentences
being considered by issue #2576, whose resolution could
'overturn' this choice, or other choices made in this PR)