Simplify PropertyDefinitionEvaluation's return type
As it stands, PDE returns "a Completion Record normally containing a Boolean",
but this is incorrect...
- the 2nd definition has `Return ? CopyDataProperties`,
  which will return an Object,
- the 4th definition has `Return ~empty~`, and
- the 5th definition has `Return ? MethodDefinitionEvaluation ...`,
  which might return a PrivateElement.

However, rather than add these possibilities into the return type,
we can actually simplify it. Note that the spec has 3 calls to PDE:
- 2 are `Perform ? PDE of ...`, and
- 1 is `Return ? PDE of ...`, but it's within the definition of PDE itself,

which means that ultimately, nobody cares about PDE's normal return.
So it might as well return ~unused~.