Editorial: Use consistent phrasing in ValidateAndApplyPropertyDescriptor (#2624)
In the "Create" and "Replace" steps in ValidateAndApplyPropertyDescriptor,
use consistent phrasing for getting a value from _Desc_ or the default.

In the "Replace" steps, clarify the "each other attribute" wording
by naming the other attributes explicitly.

And instead of talking about a record field being "present" or "absent",
use the "record has a field" phrasing that PR #2620 settled on.