Editorial: Eliminate `_NcapturingParens_` from Early Error rules + grammar (#2625)
 This PR addresses a problem similar to that of issue #1884.

`_NcapturingParens_` is one of the aliases that are defined in [22.2.2.1 Notation](https://tc39.es/ecma262/#sec-notation) and then referred to 'out of the blue' by various operations.

The general approach to such cases is to make the necessary changes to pass the value explicitly to the operations that reference it.  However, `_NcapturingParens_` is also referenced by two Early Error rules and a grammar production, and there's no way to "pass a value" to such constructs.

Instead, I've replaced those references with invocations of new abstract ops:
- TheNumberOfLeftCapturingParensIn
- TheNumberOfLeftCapturingParensInThePatternContaining

I've also defined a third AO:
- TheNumberOfLeftCapturingParensToTheLeftOf

This isn't necessary to deal with `_NcapturingParens_`, but it brings together all mentions of left-capturing parentheses and how to find them.

I believe this centralization also fixes a subtle bug in the spec...  In the main Pattern grammar, all left-capturing parentheses arise from the production:
```
    Atom :: `(` GroupSpecifier Disjunction `)`
```
But in the Annex B Pattern grammar, there are *two* productions that account for left-capturing parentheses: in a +UnicodeMode pattern, it's the same production as above, but in a ~UnicodeMode pattern, it's:
```
    ExtendedAtom :: `(` Disjunction `)`
```
So Annex B needs to generalize the main-body logic from one production to both.  It *has* a sentence to perform this generalization, but only for the CompileSubpattern defn for `Term :: Atom Quantifier`, and [this is the bug] *not* in the other places where it's relevant:
- the definition of `_NcapturingParens_`, and
- the CompileAtom defn for ```AtomEscape :: `k` GroupName```

With the logic for left-capturing parentheses centralized, it's easy for AnnexB to perform the generalization for everywhere that's relevant.

---

Yes, the AO names are pretty long. Feel free to suggest shorter ones.
- drop "The"?
- "ToTheLeftOf" -> "Before"?
- "LeftCapturingParens" -> "Groups" or "CapturingGroups"?

(Re the last, note that "TheNumberOfGroupsToTheLeftOf" would be unclear, so perhaps "TheNumberOfGroupStartsToTheLeftOf" or "TheNumberOfGroupsThatStartBefore")

---

(This PR leaves references to `_NcapturingParens_` in operations, but those can be cleaned up in a more conventional way.)