Editorial: Change GroupSpecifier from nillable to optional (#2722)
That is, change:
    Atom :: `(` GroupSpecifier Disjunction `)`
    GroupSpecifier[UnicodeMode] :: [empty] | `?` GroupName
to:
    Atom :: `(` GroupSpecifier? Disjunction `)`
    GroupSpecifier[UnicodeMode] :: `?` GroupName
(and similarly for ExtendedAtom).

This doesn't change the language, it just means that
when we have a GroupSpecifier "in hand", we know that it has a GroupName child.

(As for the case where a GroupSpecifier is empty/omitted, the spec
doesn't specifically talk about that, so no changes are necessary.)

---

Extend CapturingGroupName to GroupSpecifier

Now that GroupSpecifier is defined by a chain production,
we can apply CapturingGroupName to a GroupSpecifier,
and let the chain rule handle it.

---

Reword the early error rule involving CapturingGroupName

Currently, the spec talks about
    "multiple [things that] have the same CapturingGroupName"
which is unusual wording by which to invoke an SDO.

Change it to:
    "two or more [things] for which CapturingGroupName of [thing] is the same"
which is still unusual, but maybe less so.