Editorial: Eliminate RegExp's "global aliases" (#2716)
That is, eliminate the aliases defined in 22.2.2.1 Notation
that are magically accessible to all RegExp algorithms.

Introduce 'RegExp Record' to store information about a RegExp
that is needed during compilation and possibly during matching.
Pass it around to the various RegExp operations.

The global aliases are then eliminated in various ways:

- Each of the 'flag' aliases
  (`_DotAll_`, `_IgnoreCase_`, `_Multiline_`, `_Unicode_`)
  becomes a field of the same name in RegExp Record.

- `_NcapturingParens_` becomes the field `[[CapturingGroupsCount]]`.

- `_WordCharacters_` is replaced with on-the-fly calls to
  the (reinstated) `WordCharacters()` operation.

- `_InputLength_` is 'calculated' as needed.

- `_Input_` becomes a component of the `State` structure.